+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+## 0.0.6 - FTP Commands Support.
+
+* Add Change Directory Feature.
+* Add Quit Feature.
+* Add Type Command.
+
+
+## 0.0.5 - FTP Commands Support.
+
+* Add Server Listener to listen to client commands.
+* Add Commands Interpreter.
+* Add Welcome Message.
+* Add Username Support.
+* Add Password Support.
+
+## 0.0.4 - Server Support.
+
+* Add Welcome Message.
+
+## 0.0.3 - Server Support.
+
+* Start Server.
+
 ## 0.0.1 - Initial Release.
 
 * Add TCP Server.